Reorganize the SST cp files to separate read and write functionality, fix the read side so that we handle contact failure (like if we have an old contact file), cleanup unnecessary things in the helloSst*.cpp files
