+++ source/adios2/toolkit/sst/cp/cp_internal.h
@@ -355,3 +355,6 @@ extern void CP_WriterCloseHandler(CManager cm, CMConnection conn, void *msg_v,
 
 extern void FFSMarshalInstallMetadata(SstStream Stream, TSMetadataMsg MetaData);
 extern void FFSClearTimestepData(SstStream Stream);
+extern int *setupPeerArray(int MySize, int MyRank, int PeerSize);
+extern void CP_verbose(SstStream Stream, char *Format, ...);
+extern struct _CP_Services Svcs;
